minor structure changes
