+++ src/store/data-process/data-process.selectors.ts
@@ -1,9 +1,8 @@
 import {RootState} from '../../types/state.ts';
-import {StoreNameSpace} from '../../const.ts';
+import {SortingOrder, StoreNameSpace} from '../../const.ts';
 import {City} from '../../types/city.ts';
 import {OfferShort, OffersShort} from '../../types/offers/offer-short.ts';
 import {Nullable} from '../../types/nullable.ts';
-import {SortingOrder} from '../../types/sorting-order.ts';
 
 export const getActiveCity = (state: RootState): City => state[StoreNameSpace.Data].activeCity;
 export const getOffers = (state: RootState): OffersShort => state[StoreNameSpace.Data].offers;
