+++ src/types/sorting-order.ts
@@ -1,6 +1,1 @@
-export enum SortingOrder {
-  popular = 'Popular',
-  priceToHigh = 'Price: low to high',
-  priceToLow = 'Price: high to low',
-  topRated = 'Top rated first',
-}
+
